Custom Hook to manage Application.js
- create a custom Hook to provide the state and actions used to change the state
